Add logic to skip building source then running under Jenkins

+++ Makefile
@@ -1,6 +1,8 @@
 # XXX make sure enough space
-aaa := $(shell [ ! -f /swapfile ] && sudo dd if=/dev/zero of=/swapfile bs=1024000 count=6k)
-bbb := $(shell sudo mkswap /swapfile && sudo swapon /swapfile)
+ifeq ($(wildcard ~/.ihaveaswapfile),)
+	aaa := $(shell [ ! -f /swapfile ] && sudo dd if=/dev/zero of=/swapfile bs=1024000 count=6k)
+	bbb := $(shell sudo mkswap /swapfile && sudo swapon /swapfile)
+endif
 
 topdir         := .
 user_ext_build := true
@@ -14,7 +16,10 @@ user_build_dir := \
     hiredis \
     cryptopp \
     gmock-1.7.0 \
-    cmdconsole \
-    source
+    cmdconsole
+
+ifndef JENKINS_URL 
+    user_build_dir += source
+endif
 
 include $(topdir)/Makefile.incl
